Tinkering

+++ src/main/resources/assets/nvidium/shaders/terrain/frag.frag
@@ -14,9 +14,9 @@
 
 layout(location = 0) out vec4 colour;
 layout(location = 1) in Interpolants {
-    vec4 tint;
-    vec4 addin;
-    vec4 uv_bias_cutoff;
+    f16vec4 tint;
+    f16vec4 addin;
+    f16vec4 uv_bias_cutoff;
 };
